refactor: Change directory for AggregationsBarGraph
